integrationテストをmodelsに統合
